# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api/"
        }
    ],
    "paths": {
...
        "/v1/photos": {
            "get": {
                "tags": [
                    "Photos"
                ],
                "summary": "Get user photos",
                "description": "Get all photos for the authenticated user with optional size parameter",
                "operationId": "getUserPhotos",
                "parameters": [
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Image size preference (thumbnail, medium, original)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "medium",
                            "enum": [
                                "thumbnail",
                                "medium",
                                "original"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Photos retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "image_url": {
                                                        "type": "string",
                                                        "example": "https://example.com/storage/photos/user_1_medium.jpg"
                                                    },
                                                    "original_url": {
                                                        "type": "string",
                                                        "example": "https://example.com/storage/photos/user_1_original.jpg"
                                                    },
                                                    "medium_url": {
                                                        "type": "string",
                                                        "example": "https://example.com/storage/photos/user_1_medium.jpg"
                                                    },
                                                    "thumbnail_url": {
                                                        "type": "string",
                                                        "example": "https://example.com/storage/photos/user_1_thumbnail.jpg"
                                                    },
                                                    "is_profile_photo": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "order": {
                                                        "type": "integer",
                                                        "example": 0
                                                    },
                                                    "is_private": {
                                                        "type": "boolean",
                                                        "example": false
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-01-01T00:00:00.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-01-01T00:00:00.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}